Consistency of log base in error computation
It struck me as odd that the errors D1 and D2, which are clearly base-10 log, get compared on equal footing with D3, which is base-2 log. Is that correct? it would seem the precision is a bit overestimated here (it should hardly ever matter, because by the time D3 is dominant, it doesn't seem we'd be able to do much to get a more accurate result anyway)