Actually fix CI problem, remove workaround
The actual bug was that pip fails to properly resolve dependencies, except if it is switched to a new dependency resolver, which will become the default in the future for pip >= 20.3.

Example:
Packagage B depends on package A >= 1.0.
Assume that A version 0.9 is already installed.
Then, `pip install B` will not upgrade package A and leave B in a broken state.