Fix the comparison in test_compatibility for NumPy 2.0.0
The comparison in `mpmath/tests/test_convert.py::test_compatibility`
failed for `np.float16` in NumPy 2.0.0 since `2.0**-53` cannot be
represented in half-precision floating point type.  Explicitly convert
it to `np.float64` to ensure that the comparison is done in sufficiently
precise type.

This fixes the actual test failure from #815 but the deprecation warning
remains.